pyproject.toml
[tool.poetry]
name = "kaila_companion"
version = "0.0.1"
description = "Companion app for the Kaila TTRPG"
authors = ["Kaila-TTRPG <https://github.com/Kaila-TTRPG>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.7"
wheel = ">=0.37"
pythonnet = [
    { version = ">=2.5", platform = "win32" }
]
cefpython3 = [
    { version = ">=66.0", platform = "win32" }
]
pywebview = [
    { version = ">=3.4", platform = "darwin" },
    { version = ">=3.4", platform = "win32", extras=["cef"] },
    { version = ">=3.4", platform = "linux", extras=["qt"] }
]

[tool.poetry.dev-dependencies]
poetry = ">=1.1"
spells = ">=0.0.2"
flake8 = ">=3.8.4"
pytest = ">=6.2.1"
pytest-cov = ">=2.10"
Nuitka = ">=0.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
